requirements.txt
azure-iot-device
beautifulsoup4
lxml
python-dotenv